+++ gfx/drivers_font_renderer/bitmapfont_6x10.c
@@ -176,7 +176,7 @@ bitmapfont_lut_t *bitmapfont_6x10_load(unsigned language)
          for (i = 0; i < FONT_6X10_WIDTH; i++)
          {
             uint8_t rem     = 1 << ((i + j * FONT_6X10_WIDTH) & 7);
-            unsigned offset = (i + j * FONT_6X10_WIDTH) >> 3;
+            size_t offset   = (i + j * FONT_6X10_WIDTH) >> 3;
 
             /* LUT value is 'true' if specified glyph
              * position contains a pixel */
